fiks crash in reaction

+++ events/reaction.ts
@@ -1,4 +1,4 @@
-import { Events, MessageReaction, User } from "discord.js";
+import { Collection, Events, MessageReaction, User } from "discord.js";
 import type { BotEvent } from "../types.js";
 
 const event: BotEvent<MessageReaction> = {
@@ -13,14 +13,21 @@ const event: BotEvent<MessageReaction> = {
       }
     }
 
-    const users = await reaction.users.fetch();
+    let users = new Collection<string, User>();
+    try {
+      users = await reaction.users.fetch();
+    } catch {
+      users = reaction.users.cache;
+    }
 
-    if (
-      reaction.emoji.name === "❌" &&
-      reaction.client.user.id === reaction.message.author?.id &&
-      users.some((u) => u.id === reaction.message.interaction?.user.id)
-    ) {
-      await reaction.message.delete();
+    if (users.size > 0) {
+      if (
+        reaction.emoji.name === "❌" &&
+        reaction.client.user.id === reaction.message.author?.id &&
+        users.some((u) => u.id === reaction.message.interaction?.user.id)
+      ) {
+        await reaction.message.delete();
+      }
     }
   },
 };
